Refactor environment imports to use the correct path and remove deprecated environments file
