--- front/src/app/environments/environments.ts
@@ -1,4 +1,0 @@
-export const environment = {
-  production: false,
-  apiUrl: 'http://localhost:8080',
-};
