+++ front/src/app/shared/components/password-input/password-input.component.html
@@ -8,8 +8,8 @@
     placeholder="Wprowadź hasło"
     required
   />
-  @if (getError(passwordControl); as message) {
-    <mat-error>{{ message }}</mat-error>
+  @if (getError(passwordControl)) {
+    <mat-error>{{ getError(passwordControl) }}</mat-error>
   }
 </mat-form-field>
 
@@ -23,7 +23,7 @@
     placeholder="Wprowadź hasło"
     required
   />
-  @if (getError(repeatPasswordControl); as message) {
-    <mat-error>{{ message }}</mat-error>
+  @if (getError(repeatPasswordControl)) {
+    <mat-error>{{ getError(repeatPasswordControl) }}</mat-error>
   }
 </mat-form-field>
